# docs/openapi.yaml
components:
    schemas:
...
        dto.ErrorItem:
            properties:
                name:
                    type: string
                reason:
                    type: string
            type: object
        dto.ErrorResponse:
            properties:
                errors:
                    items:
                        $ref: '#/components/schemas/dto.ErrorItem'
                    type: array
                message:
                    type: string
            type: object
...
        dto.GetUserProfileResponse:
            properties:
                avatar:
                    type: string
                created_at:
                    type: string
                deleted_at:
                    type: string
                email:
                    type: string
                id:
                    type: string
                name:
                    type: string
                subscription_expires_at:
                    type: string
                synchronized_at:
                    type: string
                tier:
                    type: string
                updated_at:
                    type: string
            type: object
...
        dto.UpdateProfileRequest:
            properties:
                email:
                    type: string
                name:
                    type: string
            type: object
...
    securitySchemes:
        BasicAuth:
            scheme: basic
            type: http
        BearerAuth:
            description: Type "Bearer" followed by a space and JWT token.
            in: header
            name: Authorization
            type: apiKey
info:
    contact:
        email: danielmesquitta123@gmail.com
        name: Daniel Mesquita
    description: API Finance Manager
    title: API Finance Manager
    version: "1.0"
openapi: 3.0.3
paths:
...
    /v1/users/profile:
        delete:
            description: Delete logged-in user
            responses:
                "204":
                    description: No Content
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Delete logged-in user
            tags:
                - User
        get:
            description: Get logged-in user profile
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.GetUserProfileResponse'
                    description: OK
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Get logged-in user profile
            tags:
                - User
        put:
            description: Update logged-in user
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/dto.UpdateProfileRequest'
                description: Request body
                required: true
                x-originalParamName: request
            responses:
                "204":
                    description: No Content
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Update logged-in user
            tags:
                - User